Version 2.1.1
Enhanced the script to accept any regular expression as a separator pattern, but it works only for read-only operations. Now you can treat the following as an array:

axxbxxxxcxdxxxxxe

and access a, b, c, d and e as array elements.

Now the element is escaped before using in the matches, so you shouldn't have any problem using any random strings to store in to the array.

Also the script is now made as a plugin, so it works only 6.0