Version 3.5
CAUTION: This version potentially introduces an incompatibility with that of
older versions because of the below change. If your plugin depended on this
undocumented behavior you will have to update your plugin immediately to work
with the new plugin. Even if you didn't know about behavior, it is still better
to check if your plugin accidentally depended on this. If you need further
clarifications, please email me with your questions and I will try to answer.

- Now the plugin ignores the 'ignorecase' and 'smartcase' settings, so the
  results are more consistent. This I think is the right thing to do for a
  library.
- Also added a new function MvIterPeek.